requirements.txt
Flask
pandas
numpy
faiss-cpu
sentence-transformers
